Separated code and placed into the FAssetLoader class which follows the same pattern as FFilamentAsset and it's base class.
